Added NFC Functionality.
If you'd like me to move the bulk of this to a new class, let me know.
Also Todd, this is the only way I could figure out to call
BrowserActivity and pass the URL to it without breaking the listener in
DevicesFragment. If there's a better way please tell me about it cause
this kind of stumped me.